working ugly timeouts

+++ src/gameserver/src/planetwars/game_controller.rs
@@ -28,5 +28,5 @@ pub trait GameController<C: DeserializeOwned> {
     /// Returns the player id of the winner(s) or None if the game is still going
     fn outcome(& self) -> Option<Vec<usize>>;
 
-    fn time_out(& self) -> usize;
+    fn time_out(& self) -> u64;
 }
