+++ src/gameserver/src/planetwars/lock.rs
@@ -8,7 +8,7 @@ pub trait Lock<G: GameController<C>, C: DeserializeOwned> {
     /// Creates new Lock
     fn new (game_controller: G, awaiting_clients: HashSet<usize>) -> Self;
     /// Flushes the lock, maybe returning a vector of winner ids
-    fn do_step(&mut self) -> (usize, Option<Vec<usize>>);
+    fn do_step(&mut self) -> (u64, Option<Vec<usize>>);
 
     /// Returns whether or not the lock has all msgs of the clients
     /// That he is waiting for
@@ -23,5 +23,5 @@ pub trait Lock<G: GameController<C>, C: DeserializeOwned> {
     /// till it connects again
     fn disconnect(&mut self, client_id: usize);
     /// Lock act's, mainly used for timouts
-    fn do_time_out(&mut self) -> (usize, Option<Vec<usize>>);
+    fn do_time_out(&mut self) -> (u64, Option<Vec<usize>>);
 }
